refactor: refactor code according to suggestions (DD-7)
