+++ .eslintrc.json
@@ -43,7 +43,6 @@
             ]
           }
         ],
-        "no-underscore-dangle": "off",
         "@typescript-eslint/no-explicit-any": 2,
         "@typescript-eslint/explicit-member-accessibility": ["error", { "accessibility": "explicit" }],
         "import/prefer-default-export": "off",
