+++ src/app/app.component.scss
@@ -1,7 +1,4 @@
 :host {
-  max-width: 1440px;
-  margin: 0 auto;
-  padding: 0 1rem;
   display: flex;
   flex-direction: column;
 
@@ -9,11 +6,6 @@
   min-height: 100vh;
   margin: 0 auto;
   padding: 0 1rem;
-}
-
-main {
-  display: flex;
-  justify-content: center;
 }
 
 main {
