+++ src/app/features/home/services/search/search.service.ts
@@ -1,30 +1,22 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable, signal } from '@angular/core';
-import { CityInfo } from '@features/search/interfaces/city-info';
+import { CityInfo } from '@features/home/interfaces/city-info';
 import { SearchRouteParams } from '../../interfaces/search-route-params';
 
 @Injectable({
   providedIn: 'root',
 })
 export class SearchService {
-  private _cities = signal<CityInfo[]>([]);
+  private readonly _cities = signal<CityInfo[]>([]);
 
   public cities = this._cities.asReadonly();
 
   private readonly httpClient = inject(HttpClient);
 
-  public search({ fromLatitude, fromLongitude, toLatitude, toLongitude, time }: SearchRouteParams) {
-    return this.httpClient
-      .get('/api/search', {
-        params: {
-          fromLatitude,
-          fromLongitude,
-          toLatitude,
-          toLongitude,
-          time,
-        },
-      })
-      .pipe();
+  public search(params: SearchRouteParams) {
+    return this.httpClient.get('/api/search', {
+      params: { ...params },
+    });
   }
 
   public setCities(receivedCities: CityInfo[]) {
