--- src/app/features/search/components/search-form/search-form.component.scss
@@ -1,9 +1,0 @@
-.container {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.search-input {
-  min-width: 20%;
-}
