+++ src/environment.ts
@@ -1,4 +1,4 @@
 export const environment = {
   baseLocationApiUrl: 'http://api.openweathermap.org/geo/1.0',
   locationApiKey: import.meta.env.NG_APP_API_LOCATION_KEY,
-};
+} as const;
